Use std futures, upgrade wasm-bindgen to latest
