+++ examples/server_integration/client/src/example_b.rs
@@ -1,4 +1,4 @@
-use seed::fetch;
+use seed::browser::service::fetch;
 use seed::prelude::*;
 use serde::Deserialize;
 use std::borrow::Cow;
