+++ examples/update_from_js/src/lib.rs
@@ -1,5 +1,5 @@
 use enclose::enc;
-use seed::prelude::*;
+use seed::{prelude::*, *};
 
 // Model
