+++ examples/user_media/src/lib.rs
@@ -1,5 +1,5 @@
-use seed::prelude::*;
 use seed::{document, window};
+use seed::{prelude::*, *};
 use wasm_bindgen::JsCast;
 use wasm_bindgen_futures::JsFuture;
 use web_sys::{HtmlMediaElement, MediaStream, MediaStreamConstraints};
