+++ src/app.rs
@@ -11,7 +11,8 @@ use builder::{
     IntoAfterMount, MountPointInitInitAPI, UndefinedInitAPI, UndefinedMountPoint,
 };
 use enclose::enclose;
-use futures::Future;
+use futures::future::LocalFutureObj;
+use futures::FutureExt;
 use std::{
     cell::{Cell, RefCell},
     collections::VecDeque,
@@ -337,29 +338,27 @@ impl<Ms, Mdl, ElC: View<Ms> + 'static, GMs: 'static> App<Ms, Mdl, ElC, GMs> {
         orders.effects
     }
 
-    fn process_queue_cmd(&self, cmd: Box<dyn Future<Item = Ms, Error = Ms>>) {
+    fn process_queue_cmd(&self, cmd: LocalFutureObj<'static, Result<Ms, Ms>>) {
         let lazy_schedule_cmd = enclose!((self => s) move |_| {
             // schedule future (cmd) to be executed
-            spawn_local(cmd.then(move |res| {
-                let msg_returned_from_effect = res.unwrap_or_else(|err_msg| err_msg);
+            spawn_local(async move {
+                let msg_returned_from_effect = cmd.await.unwrap_or_else(|err_msg| err_msg);
                 // recursive call which can blow the call stack
                 s.update(msg_returned_from_effect);
-                Ok(())
-            }))
+            })
         });
         // we need to clear the call stack by NextTick so we don't exceed it's capacity
         spawn_local(NextTick::new().map(lazy_schedule_cmd));
     }
 
-    fn process_queue_global_cmd(&self, g_cmd: Box<dyn Future<Item = GMs, Error = GMs>>) {
+    fn process_queue_global_cmd(&self, g_cmd: LocalFutureObj<'static, Result<GMs, GMs>>) {
         let lazy_schedule_cmd = enclose!((self => s) move |_| {
             // schedule future (g_cmd) to be executed
-            spawn_local(g_cmd.then(move |res| {
-                let msg_returned_from_effect = res.unwrap_or_else(|err_msg| err_msg);
+            spawn_local(async move {
+                let msg_returned_from_effect = g_cmd.await.unwrap_or_else(|err_msg| err_msg);
                 // recursive call which can blow the call stack
                 s.sink(msg_returned_from_effect);
-                Ok(())
-            }))
+            })
         });
         // we need to clear the call stack by NextTick so we don't exceed it's capacity
         spawn_local(NextTick::new().map(lazy_schedule_cmd));
